fixed random modal bug
